airflow_dag/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages

astronomer-cosmos
apache-airflow-providers-snowflake==4.0.2
astro-sdk-python [amazon,snowflake] >= 1.1.0